+++ src/headers/DirectionalLight.h
@@ -8,7 +8,7 @@ public:
     DirectionalLight();
     DirectionalLight(GLfloat red, GLfloat green, GLfloat blue, GLfloat ambientIntensity, GLfloat xDir, GLfloat yDir, GLfloat zDir, GLfloat diffuseIntensity);
 
-    void UseLight(GLfloat ambientIntesityLocation, GLfloat ambienColourLocation, GLfloat diffuseIntensityLocation, GLfloat directionLocation);
+    void UseLight(GLuint ambientIntesityLocation, GLuint ambienColourLocation, GLuint diffuseIntensityLocation, GLuint directionLocation);
 
     ~DirectionalLight();
